rag-python/requirements.txt
# rag-python/requirements.txt

# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic-settings==2.2.1

# RAG & LLM
langchain==0.2.1
langchain-community==0.2.1
sentence-transformers==3.0.1
langchain-google-genai==1.0.6 # GeminiモデルをLangChainでより簡単に扱うための推奨パッケージ
chromadb==0.5.0 # google-generativeaiはlangchain-google-genaiが依存関係としてインストールするため削除

# Authentication
PyJWT==2.8.0
cryptography==42.0.7 # PyJWTの依存関係

# File Processing
pypdf==4.2.0
python-docx==1.1.2
python-multipart==0.0.9

# Utilities
python-dotenv==1.0.1
werkzeug==3.0.3
sentencepiece==0.2.0 # 高速トークナイザーのために必要